pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=2.16.0",
    "firebase-admin>=7.1.0",
    "google-api-python-client>=2.182.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-firestore>=2.21.0",
    "google-genai>=1.38.0",
    "notion-client>=2.5.0",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "streamlit>=1.49.1",
    "twilio>=9.8.1",
]
